fix: resolve comments, reuse time functions, split templates into files
